website

+++ app/Http/Controllers/AuthController.php
@@ -34,7 +34,10 @@ class AuthController extends Controller
     return redirect()->back()->with('message','بيانات خاطئه');
   }
   public function Logout()
-  {
+  {if (Auth::user()->role('مقدم خدمه')||Auth::user()->role('طالب خدمه')) {
+    Auth::logout();
+    return redirect('Login');
+  }
     Auth::logout();
     return redirect('AdminLogin');
   }
@@ -53,7 +56,7 @@ class AuthController extends Controller
       if ($this->checkforget($user)) {
         $forgetPassowrd=forgetPassowrd::where('email', $user->email)->first();
         $forgetPassowrd->email=$user->email;
-        $forgetPassowrd->token=$this->getToken(60);;
+        $forgetPassowrd->token=mt_rand(10000000,99999999);
         $forgetPassowrd->save();
 
       }
@@ -61,35 +64,44 @@ class AuthController extends Controller
 
         $forgetPassowrd=new forgetPassowrd();
         $forgetPassowrd->email=$user->email;
-        $forgetPassowrd->token=$this->getToken(60);;
+        $forgetPassowrd->token=mt_rand(10000000,99999999);
         $forgetPassowrd->save();
       }
       //$user->notify(new ResetPassord($user));
       dispatch(new ResetPasswordJob($user));
-      return redirect('AdminLogin')->with('success_message','تم ارسال الرابط تفقد البريد الخاص بك');
+      return redirect('CodeValidation')->with('success_message','تم ارسال الرابط تفقد البريد الخاص بك');
     }
-    return redirect('ResetPassord')->with('message','هذا الاميل غير موجود ');
+    return redirect('ResetPassord')->with('message','هذا البريد الالكتروني غير موجود ');
   }
 
-  public function NewPassword($token,$email)
+  public function NewPassword(request $token)
   {
-    $forgetPassowrd=forgetPassowrd::where('token',$token)->first();
+    $forgetPassowrd=forgetPassowrd::where('token',$token->token)->first();
     if ($forgetPassowrd) {
-
+      $email=$forgetPassowrd->email;
+      $token=$forgetPassowrd->token;
       return view('admin/Auth/NewPassword',compact('token','email'));
     }
     else {
-      return redirect('AdminLogin');
+      return redirect('Login')->with('message','هذا الكود خاطئ ');
     }
   }
 
 
- public function ResetNewPassword(newpassrequest $data){
+ public function ResetNewPassword(Request $data){
+
+if (strlen($data->password)<6||($data->password)!=($data->password_confirmation)) {
+  return redirect('Login')->with('message','تأكد من كلمة المرور جيدا ');
+}
+
     $user=user::where('email',$data->email)->first();
     $user->password=bcrypt($data->password);
     $user->save();
     $forgetPassowrd=forgetPassowrd::where('token',$data->token)->first()->delete();
-    return redirect('AdminLogin');;
+    if ($user->role=='مقدم خدمه'||$user->role=='طالب خدمه') {
+      return redirect('Login')->with('message','تم تغير كلمة السر بنجاح');
+    }
+    return redirect('AdminLogin')->with('message','تم تغير كلمه السر بنجاح');
  }
 
   public function checkforget($user)
@@ -116,5 +128,10 @@ class AuthController extends Controller
     }
 
     return $token;
+  }
+
+  public function CodeValidation()
+  {
+    return view('admin\Auth\CodeValidation');
   }
 }
